@@ -345,26 +345,33 @@ namespace Test
         public static void TestLibraryGeneration()
         {
             string thisTaskOutputFolder = Path.Combine(TestContext.CurrentContext.TestDirectory, @"TestData\SpectralLibrarySearch\FileOutput");
+            _ = Directory.CreateDirectory(thisTaskOutputFolder);
 
             SearchTask task = Toml.ReadFile<SearchTask>(Path.Combine(TestContext.CurrentContext.TestDirectory, @"TestData\SpectralLibrarySearch\SpectralSearchTask.toml"), MetaMorpheusTask.tomlConfig);
             task.SearchParameters.WriteMzId = true;
             task.SearchParameters.WriteSpectralLibrary = true;
 
-            DbForTask db = new DbForTask(Path.Combine(TestContext.CurrentContext.TestDirectory, @"TestData\SpectralLibrarySearch\uniprot-yeast-filtered-reviewed_yes.fasta.gz"), false);
-            string raw = Path.Combine(TestContext.CurrentContext.TestDirectory, @"TestData\SpectralLibrarySearch\slicedYeastData.raw");
-            EverythingRunnerEngine MassSpectraFile = new EverythingRunnerEngine(new List<(string, MetaMorpheusTask)> { ("SpectraFileOutput", task) }, new List<string> { raw }, new List<DbForTask> { db }, thisTaskOutputFolder);
+            DbForTask db = new(Path.Combine(TestContext.CurrentContext.TestDirectory, @"TestData\hela_snip_for_unitTest.fasta"),false);
+
+            string raw = Path.Combine(TestContext.CurrentContext.TestDirectory, @"TestData\TaGe_SA_HeLa_04_subset_longestSeq.mzML");
+
+            //we'll make a copy so that when we search this copy, we get enough psms to compute pep q-value. It gets deleted below.
+            string rawCopy = Path.Combine(TestContext.CurrentContext.TestDirectory, @"TestData\SpectralLibrarySearch\FileOutput\rawCopy.mzML");
+            File.Copy(raw, rawCopy);
+
+            EverythingRunnerEngine MassSpectraFile = new(new List<(string, MetaMorpheusTask)> { ("SpectraFileOutput", task) }, new List<string> { raw, rawCopy }, new List<DbForTask> { db }, thisTaskOutputFolder);
 
             MassSpectraFile.Run();
-            var test = Path.Combine(thisTaskOutputFolder, @"SpectraFileOutput\spectralLibrary.msp");
+            string test = Path.Combine(thisTaskOutputFolder, @"SpectraFileOutput\spectralLibrary.msp");
 
-            var testDir = Path.Combine(TestContext.CurrentContext.TestDirectory, @"TestData\SpectralLibraryGenaration");
-            var outputDir = Path.Combine(testDir, @"SpectralLibraryTest");
+            string testDir = Path.Combine(TestContext.CurrentContext.TestDirectory, @"TestData\SpectralLibraryGenaration");
+            string outputDir = Path.Combine(testDir, @"SpectralLibraryTest");
 
-            Directory.CreateDirectory(outputDir);
+            _ = Directory.CreateDirectory(outputDir);
 
-            var searchTask = new SearchTask();
+            SearchTask searchTask = new();
 
-            searchTask.RunTask(outputDir,
+            _ = searchTask.RunTask(outputDir,
                 new List<DbForTask>
                 {
                     new DbForTask(test, false),
@@ -373,18 +380,15 @@ namespace Test
                 new List<string> { raw },
                 "");
 
-            var results = File.ReadAllLines(Path.Combine(outputDir, @"AllPSMs.psmtsv"));
-            var split = results[0].Split('\t');
+            string[] results = File.ReadAllLines(Path.Combine(outputDir, @"AllPSMs.psmtsv"));
+            string[] split = results[0].Split('\t');
             int ind = Array.IndexOf(split, "Normalized Spectral Angle");
             int indOfTarget = Array.IndexOf(split, "Decoy/Contaminant/Target");
             Assert.That(ind >= 0);
-            var spectralAngleList = new List<Double>();
-            var decoySpectralAngleList = new List<Double>();
+            List<double> spectralAngleList = new();
             for (int i = 1; i < results.Length; i++)
             {
-                String sequence = results[i].Split('\t')[14].ToString();
-
-                var spectralAngle = double.Parse(results[i].Split('\t')[ind]);
+                double spectralAngle = double.Parse(results[i].Split('\t')[ind]);
                 string targetOrDecoy = results[i].Split('\t')[indOfTarget].ToString();
 
                 if (targetOrDecoy.Equals("T") && spectralAngle >= 0)
@@ -415,39 +419,6 @@ namespace Test
             Assert.That(decoySpectum[1].NeutralTheoreticalProduct.ProductType == ProductType.b && decoySpectum[1].NeutralTheoreticalProduct.FragmentNumber == 2 && decoySpectum[1].Intensity == 2);
             Assert.That(decoySpectum[2].NeutralTheoreticalProduct.ProductType == ProductType.b && decoySpectum[2].NeutralTheoreticalProduct.FragmentNumber == 3 && decoySpectum[2].Intensity == 3);
             Assert.That(decoySpectum[3].NeutralTheoreticalProduct.ProductType == ProductType.b && decoySpectum[3].NeutralTheoreticalProduct.FragmentNumber == 4 && decoySpectum[3].Intensity == 4);
-        }
-
-        [Test]
-        public static void TestDecoyLibraryGenerationDuringSearch()
-        {
-            string thisTaskOutputFolder = Path.Combine(TestContext.CurrentContext.TestDirectory, @"TestData\SpectralLibrarySearch\FileOutput");
-
-            var task = new SearchTask();
-
-            var spectralLibrary = Path.Combine(TestContext.CurrentContext.TestDirectory, @"TestData\SpectralLibrarySearch\EPAVFGR.msp");
-            DbForTask db = new DbForTask(Path.Combine(TestContext.CurrentContext.TestDirectory, @"TestData\SpectralLibrarySearch\mouse-filtered-reviewed.fasta.gz"), false);
-            List<DbForTask> dbList = new List<DbForTask> { new DbForTask(spectralLibrary, false), db };
-            string raw = Path.Combine(TestContext.CurrentContext.TestDirectory, @"TestData\SpectralLibrarySearch\slicedMouse300-480.raw");
-            EverythingRunnerEngine MassSpectraFile = new EverythingRunnerEngine(new List<(string, MetaMorpheusTask)> { ("SpectraFileOutput", task) }, new List<string> { raw }, dbList, thisTaskOutputFolder);
-
-            MassSpectraFile.Run();
-
-            var results = File.ReadAllLines(Path.Combine(thisTaskOutputFolder, @"SpectraFileOutput\AllPSMs.psmtsv"));
-            var split = results[0].Split('\t');
-            int indOfSA = Array.IndexOf(split, "Normalized Spectral Angle");
-            int indOfTorD = Array.IndexOf(split, "Decoy/Contaminant/Target");
-            int indOfTargetSequenceInTDpairs = Array.IndexOf(split, "Peptide Description");
-            int indOfsequence = Array.IndexOf(split, "Full Sequence");
-            for (int i = 1; i < results.Length; i++)
-            {
-                if (results[i].Split('\t')[indOfsequence] == "GFVAPER")
-                {
-                    Assert.That(results[i].Split('\t')[indOfTorD] == "D");
-                    Assert.That(Double.Parse(results[i].Split('\t')[indOfSA]) < 0.45);
-                    Assert.That(results[i].Split('\t')[indOfTargetSequenceInTDpairs] == "EPAVFGR");
-                }
-            }
-            Directory.Delete(thisTaskOutputFolder, true);
         }
     }
 }
