@@ -156,15 +156,15 @@ namespace Test
         {
             // create the unit test directory
             string unitTestFolder = Path.Combine(TestContext.CurrentContext.TestDirectory, @"TestProteinQuantFileHeaders");
-            Directory.CreateDirectory(unitTestFolder);
+            _ = Directory.CreateDirectory(unitTestFolder);
 
-            List<SpectraFileInfo> fileInfos = new List<SpectraFileInfo>();
+            List<SpectraFileInfo> fileInfos = new();
             string peptide = "PEPTIDE";
             double ionIntensity = 1e6;
             string condition = hasDefinedExperimentalDesign ? "TestCondition" : "";
 
             // create the protein database
-            Protein prot = new Protein(peptide, @"");
+            Protein prot = new(peptide, @"");
             string dbName = Path.Combine(unitTestFolder, "testDB.fasta");
             UsefulProteomicsDatabases.ProteinDbWriter.WriteFastaDatabase(new List<Protein> { prot }, dbName, ">");
 
@@ -184,14 +184,14 @@ namespace Test
                         ChemicalFormula cf = new Proteomics.AminoAcidPolymer.Peptide(peptide).GetChemicalFormula();
                         IsotopicDistribution dist = IsotopicDistribution.GetDistribution(cf, 0.125, 1e-8);
                         double[] mz = dist.Masses.Select(v => v.ToMz(1)).ToArray();
-                        double[] intensities = dist.Intensities.Select(v => v * ionIntensity).ToArray();
+                        double[] intensities = dist.Intensities.Select(v => v * ionIntensity * (b + 1)).ToArray();
 
                         scans[0] = new MsDataScan(massSpectrum: new MzSpectrum(mz, intensities, false), oneBasedScanNumber: 1, msnOrder: 1, isCentroid: true,
                             polarity: Polarity.Positive, retentionTime: 1.0, scanWindowRange: new MzRange(400, 1600), scanFilter: "f",
                             mzAnalyzer: MZAnalyzerType.Orbitrap, totalIonCurrent: intensities.Sum(), injectionTime: 1.0, noiseData: null, nativeId: "scan=1");
 
                         // create the MS2 scan
-                        var pep = new PeptideWithSetModifications(peptide, new Dictionary<string, Proteomics.Modification>());
+                        PeptideWithSetModifications pep = new(peptide, new Dictionary<string, Proteomics.Modification>());
                         List<Product> frags = new List<Product>();
                         pep.Fragment(DissociationType.HCD, FragmentationTerminus.Both, frags);
                         double[] mz2 = frags.Select(v => v.NeutralMass.ToMz(1)).ToArray();
@@ -209,7 +209,7 @@ namespace Test
                             new MsDataFile(scans, new SourceFile(@"scan number only nativeID format", "mzML format", null, "SHA-1", @"C:\fake.mzML", null)),
                             fullPath, false);
 
-                        var spectraFileInfo = new SpectraFileInfo(fullPath, condition, b, r, f);
+                        SpectraFileInfo spectraFileInfo = new(fullPath, condition, b, r, f);
                         fileInfos.Add(spectraFileInfo);
                     }
                 }
@@ -218,20 +218,20 @@ namespace Test
             // write the experimental design for this quantification test
             if (hasDefinedExperimentalDesign)
             {
-                ExperimentalDesign.WriteExperimentalDesignToFile(fileInfos);
+                _ = ExperimentalDesign.WriteExperimentalDesignToFile(fileInfos);
             }
 
             // run the search/quantification
             SearchTask task = new SearchTask();
-            task.RunTask(unitTestFolder, new List<DbForTask> { new DbForTask(dbName, false) }, fileInfos.Select(p => p.FullFilePathWithExtension).ToList(), "");
+            _ = task.RunTask(unitTestFolder, new List<DbForTask> { new DbForTask(dbName, false) }, fileInfos.Select(p => p.FullFilePathWithExtension).ToList(), "");
 
             // read in the protein quant results
             Assert.That(File.Exists(Path.Combine(unitTestFolder, "AllQuantifiedProteinGroups.tsv")));
-            var lines = File.ReadAllLines(Path.Combine(unitTestFolder, "AllQuantifiedProteinGroups.tsv"));
+            string[] lines = File.ReadAllLines(Path.Combine(unitTestFolder, "AllQuantifiedProteinGroups.tsv"));
 
             // check the intensity column headers
-            var splitHeader = lines[0].Split(new char[] { '\t' }).ToList();
-            var intensityColumnHeaders = splitHeader.Where(p => p.Contains("Intensity", StringComparison.OrdinalIgnoreCase)).ToList();
+            List<string> splitHeader = lines[0].Split(new char[] { '\t' }).ToList();
+            List<string> intensityColumnHeaders = splitHeader.Where(p => p.Contains("Intensity", StringComparison.OrdinalIgnoreCase)).ToList();
 
             Assert.That(intensityColumnHeaders.Count == 2);
 
@@ -254,7 +254,7 @@ namespace Test
 
             Assert.That(intensity1 > 0);
             Assert.That(intensity2 > 0);
-            Assert.That(intensity1 == intensity2);
+            Assert.That(intensity1 < intensity2);
 
             Directory.Delete(unitTestFolder, true);
         }
