@@ -1,14 +1,13 @@
 ﻿using Chemistry;
 using MassSpectrometry;
-using MassSpectrometry.MzSpectra;
-using MzLibUtil;
 using Proteomics.Fragmentation;
-using Proteomics.ProteolyticDigestion;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
+using Proteomics.ProteolyticDigestion;
 using System.Threading.Tasks;
+using Easy.Common.Extensions;
+using MassSpectrometry.MzSpectra;
 
 namespace EngineLayer
 {
@@ -61,7 +60,7 @@ namespace EngineLayer
                                         var decoyPeptideTheorProducts = new List<Product>();
                                         Peptide.Fragment(commonParameters.DissociationType, commonParameters.DigestionParams.FragmentationTerminus, decoyPeptideTheorProducts);
                                         var decoylibrarySpectrum = GetDecoyLibrarySpectrumFromTargetByReverse(targetlibrarySpectrum, decoyPeptideTheorProducts);
-                                        SpectralSimilarity s = new SpectralSimilarity(scan.TheScan.MassSpectrum, decoylibrarySpectrum.Select(x => x.Mz).ToArray(), decoylibrarySpectrum.Select(x => x.Intensity).ToArray(), SpectralSimilarity.SpectrumNormalizationScheme.squareRootSpectrumSum, commonParameters.ProductMassTolerance.Value, false);
+                                        SpectralSimilarity s = new SpectralSimilarity(scan.TheScan.MassSpectrum, decoylibrarySpectrum.Select(x => x.Mz).ToArray(),decoylibrarySpectrum.Select(x => x.Intensity).ToArray(), SpectralSimilarity.SpectrumNormalizationScheme.squareRootSpectrumSum, commonParameters.ProductMassTolerance.Value, false);
                                         if (s.SpectralContrastAngle().HasValue)
                                         {
                                             pwsms.Add((Notch, Peptide));
@@ -104,5 +103,4 @@ namespace EngineLayer
             return decoyFragmentIons;
         }
     }
-
 }
