@@ -79,7 +79,7 @@ namespace Test
             string outputFolder = Path.Combine(TestContext.CurrentContext.TestDirectory, @"TestMbrAnalysisOutput");
 
             FileSpecificParameters[] fileSpecificParameters = new FileSpecificParameters[1];
-            var filePath = Path.Combine(TestContext.CurrentContext.TestDirectory, "TestData\\SearchTaskTester.toml");
+            var filePath = Path.Combine(TestContext.CurrentContext.TestDirectory, @"TestData\FileSpecificActual.toml");
             var fileSpecificToml = Toml.ReadFile(filePath, MetaMorpheusTask.tomlConfig);
             fileSpecificParameters[0] = new FileSpecificParameters(fileSpecificToml);
 
